+++ src/UI/SliderButtons/SliderButtons.module.sass
