+++ src/components/TopTen/TopTen.module.sass
@@ -17,7 +17,7 @@
 
 .row
     margin: 0 -20px
-    padding: 20px
+    padding: 24px 20px
     overflow: hidden
 
 .swiper
