+++ src/components/TopTen/TopTenSlide/TopTenSlide.module.sass
@@ -26,7 +26,7 @@
             height: 270px
             width: 100%
             position: absolute
-            bottom: 0px
+            bottom: -1px
             z-index: 2
 
     &__img
