+++ src/types/IBannerSlide.ts
